Moved out spell enum functions
