+++ src/Engine/Objects/Character.cpp
@@ -7,6 +7,7 @@
 #include "Engine/AssetsManager.h"
 #include "Engine/Spells/CastSpellInfo.h"
 #include "Engine/Spells/Spells.h"
+#include "Engine/Spells/SpellEnumFunctions.h"
 #include "Engine/Graphics/DecalBuilder.h"
 #include "Engine/Graphics/Indoor.h"
 #include "Engine/Graphics/Image.h"
