+++ src/Engine/Spells/SpellEnums.h
@@ -1,10 +1,6 @@
 #pragma once
 
 #include <cstdint>
-#include <cassert>
-
-#include "Utility/Workaround/ToUnderlying.h"
-#include "Utility/Segment.h"
 
 enum class SpellId {
     SPELL_NONE = 0,
@@ -135,30 +131,6 @@ enum class SpellId {
 using enum SpellId;
 
 /**
- * @return                              All regular spell types.
- */
-inline Segment<SpellId> allRegularSpells() {
-    return {SPELL_FIRST_REGULAR, SPELL_LAST_REGULAR};
-}
-
-/**
- * Is spell target is item in inventory?
- */
-inline bool isSpellTargetsItem(SpellId uSpellID) {
-    return uSpellID == SPELL_WATER_ENCHANT_ITEM ||
-           uSpellID == SPELL_FIRE_FIRE_AURA ||
-           uSpellID == SPELL_DARK_VAMPIRIC_WEAPON ||
-           uSpellID == SPELL_WATER_RECHARGE_ITEM;
-}
-
-/**
- * Is spell ID references any regular spell?
- */
-inline bool isRegularSpell(SpellId uSpellID) {
-    return uSpellID >= SPELL_FIRST_REGULAR && uSpellID <= SPELL_LAST_REGULAR;
-}
-
-/**
  * Magic school, note that order corresponds to the enum order in `SPELL_TYPE`.
  */
 enum class MagicSchool {
@@ -176,24 +148,3 @@ enum class MagicSchool {
     MAGIC_SCHOOL_LAST = MAGIC_SCHOOL_DARK
 };
 using enum MagicSchool;
-
-inline Segment<MagicSchool> allMagicSchools() {
-    return {MAGIC_SCHOOL_FIRST, MAGIC_SCHOOL_LAST};
-}
-
-inline Segment<SpellId> spellsForMagicSchool(MagicSchool school) {
-    int first = 1 + std::to_underlying(school) * 11;
-    int last = first + 10;
-    return {static_cast<SpellId>(first), static_cast<SpellId>(last)};
-}
-
-inline MagicSchool magicSchoolForSpell(SpellId spell) {
-    assert(spell >= SPELL_FIRST_REGULAR && spell <= SPELL_LAST_REGULAR);
-
-    return static_cast<MagicSchool>((std::to_underlying(spell) - 1) / 11);
-}
-
-// TODO(captainurist): I think we can drop most usages of this function.
-inline int spellIndexInMagicSchool(SpellId spell) {
-    return (std::to_underlying(spell) - 1) % 11;
-}
