+++ src/GUI/UI/UISpellbook.cpp
@@ -10,6 +10,7 @@
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/Graphics/Image.h"
 #include "Engine/Spells/Spells.h"
+#include "Engine/Spells/SpellEnumFunctions.h"
 #include "Engine/Localization.h"
 #include "Engine/Party.h"
 #include "Engine/Time.h"
